Fix indentation and formatting in test_aligncore_parser
